bidesigner/pyproject.toml
[project]
name = "allauth-headless"
version = "0.1.0"
description = ""
authors = [
    {name = "lmbelo",email = "lucas.belo@live.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-allauth[socialaccount] (>=65.9.0,<66.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "dj-database-url (>=3.0.0,<4.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
